requirements.txt
# MedScribe Rural — Python Dependencies

# Web framework & server
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0

# UI
gradio==4.31.0

# OCR
paddlepaddle==2.6.1
paddleocr==2.7.3

# Image processing
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0

# Report generation
reportlab==4.2.0
pandas==2.2.2

# Database
# sqlite3 is built-in to Python

# Utilities
python-multipart==0.0.9
python-dotenv==1.0.1
